move MutableData config types into @jpmorganchase/mosaic-types
