+++ packages/types/src/MutableData.ts
@@ -1,3 +1,23 @@
+export interface Aliases {
+  [key: string]: Set<string>;
+}
+export interface FSRef {
+  $$path?: string | string[];
+  $$value?: string;
+}
+export interface GlobalRefs {
+  [key: string]: FSRef;
+}
+export interface ScopedRefs {
+  [key: string]: FSRef;
+}
+
+export interface BaseData {
+  refs?: ScopedRefs | [];
+  globalRefs?: GlobalRefs | [];
+  aliases?: Aliases | [];
+}
+
 export type MutableData<T = {}> = ImmutableData<T> & {
   /**
    * Set arbitrary data onto the config object. This value will persist across plugins and
